agentic-ai/pyproject.toml
[project]
name = "kongsberg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain",
    "langchain-openai",
    "langchain-community",
    "langgraph",
    "langchainhub",
    "tiktoken",
    "python-dotenv",
    "duckduckgo-search",
    "ddgs>=9.10.0",
]
